movie-poster-logo-placement-v2/requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==24.3.25
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
nodeenv==1.8.0
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
openai==1.30.1
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
prisma==0.13.1
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
PyMatting==1.1.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rembg==2.0.56
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.34.0
sympy==1.12
tenacity==8.3.0
tifffile==2024.5.22
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
